refactor: 모달 스타일 export 변경 / 스크롤 탑 버튼 리팩토링
