+++ src/components/Major/MajorSelect.jsx
@@ -1,10 +1,10 @@
 import { useSearchParams } from 'react-router-dom';
-import { Arrows, OptionBox, SelectedOption, SelectedOption_M } from '../styles/Common';
+import { Arrows, OptionBox, SelectedOption, SelectedOption_M } from '../../styles/Common';
 import Modal from 'react-modal';
 import { useState } from 'react';
-import { MajorModalStyle } from './Etc/ModalStyle';
-import MajorSearch from './MajorSearch';
-import { subStr } from '../utils/subString';
+import { MajorModalStyle } from '../Etc/ModalStyle';
+import MajorSearch from '../Major/MajorSearch';
+import { subStr } from '../../utils/subString';
 
 const MajorSelect = () => {
   const [searchParams] = useSearchParams();
